prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	long int n, pf;
+	int n, pf;
 
 	n = 612852475143;
 	for(pf = 2; pf <= n; pf++)
